+++ app/controllers/projects_controller.rb
@@ -267,7 +267,7 @@ class ProjectsController < ApplicationController
 
   def project_params
     params.require(:project).permit(
-      :name, :slug, :description, :acrostic_enabled, :subject_code_name,
+      :name, :slug, :description, :subject_code_name,
       :show_contacts, :show_documents, :show_posts, :disable_all_emails,
       :collect_email_on_surveys, :lockable, :hide_values_on_pdfs,
       :double_data_entry, :randomizations_enabled, :adverse_events_enabled,