+++ app/controllers/subjects_controller.rb
@@ -242,9 +242,7 @@ class SubjectsController < ApplicationController
     params[:subject][:subject_code] = params[:subject][:subject_code].strip unless params[:subject][:subject_code].blank?
     params[:subject][:site_id] = (current_user.all_editable_sites.pluck(:id).include?(params[:site_id].to_i) ? params[:site_id].to_i : nil)
     params[:subject][:project_id] = @project.id
-    params.require(:subject).permit(
-      :project_id, :subject_code, :site_id, :acrostic
-    )
+    params.require(:subject).permit(:project_id, :subject_code, :site_id)
   end
 
   def sheet_params