+++ db/migrate/20160210181731_remove_acrostics.rb
@@ -1,0 +1,6 @@
+class RemoveAcrostics < ActiveRecord::Migration
+  def change
+    remove_column :subjects, :acrostic, :string, null: false, default: ''
+    remove_column :projects, :acrostic_enabled, :boolean, null: false, default: false
+  end
+end