+++ test/controllers/subjects_controller_test.rb
@@ -297,7 +297,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should create subject' do
     assert_difference('Subject.count') do
-      post :create, project_id: @project, subject: { subject_code: 'Code03', acrostic: '', status: @subject.status }, site_id: @subject.site_id
+      post :create, project_id: @project, subject: { subject_code: 'Code03' }, site_id: @subject.site_id
     end
 
     assert_redirected_to project_subject_path(assigns(:subject).project, assigns(:subject))
@@ -305,7 +305,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should create subject and strip whitespace' do
     assert_difference('Subject.count') do
-      post :create, project_id: @project, subject: { subject_code: ' Code04 ', acrostic: '', status: @subject.status }, site_id: @subject.site_id
+      post :create, project_id: @project, subject: { subject_code: ' Code04 ' }, site_id: @subject.site_id
     end
 
     assert_equal 'Code04', assigns(:subject).subject_code
@@ -315,7 +315,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should create subject with valid subject code' do
     assert_difference('Subject.count') do
-      post :create, project_id: @project, subject: { subject_code: 'S100', acrostic: '', status: @subject.status }, site_id: sites(:site_with_subject_regex)
+      post :create, project_id: @project, subject: { subject_code: 'S100' }, site_id: sites(:site_with_subject_regex)
     end
 
     assert_redirected_to project_subject_path(assigns(:subject).project, assigns(:subject))
@@ -323,7 +323,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should not create subject with invalid subject code format' do
     assert_difference('Subject.count', 0) do
-      post :create, project_id: @project, subject: { subject_code: 'S100a', acrostic: '', status: @subject.status }, site_id: sites(:site_with_subject_regex)
+      post :create, project_id: @project, subject: { subject_code: 'S100a' }, site_id: sites(:site_with_subject_regex)
     end
 
     assert_not_nil assigns(:subject)
@@ -334,7 +334,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should not create subject with blank subject code' do
     assert_difference('Subject.count', 0) do
-      post :create, project_id: @project, subject: { subject_code: '', acrostic: '', status: @subject.status }, site_id: @subject.site_id
+      post :create, project_id: @project, subject: { subject_code: '' }, site_id: @subject.site_id
     end
 
     assert_not_nil assigns(:subject)
@@ -345,7 +345,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should not create subject with a subject code that differs only in upper or lower case to an existing subject code' do
     assert_difference('Subject.count', 0) do
-      post :create, project_id: @project, subject: { subject_code: 'code01', acrostic: '', status: @subject.status }, site_id: @subject.site_id
+      post :create, project_id: @project, subject: { subject_code: 'code01' }, site_id: @subject.site_id
     end
 
     assert_not_nil assigns(:subject)
@@ -356,7 +356,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should not create subject for invalid project' do
     assert_difference('Subject.count', 0) do
-      post :create, project_id: projects(:four), subject: { subject_code: 'Code03', acrostic: '', status: @subject.status }, site_id: @subject.site_id
+      post :create, project_id: projects(:four), subject: { subject_code: 'Code03' }, site_id: @subject.site_id
     end
 
     assert_nil assigns(:project)
@@ -368,7 +368,7 @@ class SubjectsControllerTest < ActionController::TestCase
   test 'should not create subject for site viewer' do
     login(users(:site_one_viewer))
     assert_difference('Subject.count', 0) do
-      post :create, project_id: @project, subject: { subject_code: 'Code03', acrostic: '', status: 'valid' }, site_id: sites(:one).id
+      post :create, project_id: @project, subject: { subject_code: 'Code03' }, site_id: sites(:one).id
     end
 
     assert_nil assigns(:project)
@@ -380,7 +380,7 @@ class SubjectsControllerTest < ActionController::TestCase
   test 'should create subject for site editor' do
     login(users(:site_one_editor))
     assert_difference('Subject.count') do
-      post :create, project_id: @project, subject: { subject_code: 'Code03', acrostic: '', status: 'valid' }, site_id: sites(:one).id
+      post :create, project_id: @project, subject: { subject_code: 'Code03' }, site_id: sites(:one).id
     end
 
     assert_redirected_to project_subject_path(assigns(:subject).project, assigns(:subject))
@@ -389,7 +389,7 @@ class SubjectsControllerTest < ActionController::TestCase
   test 'should not create subject for site editor for another site' do
     login(users(:site_one_editor))
     assert_difference('Subject.count', 0) do
-      post :create, project_id: @project, subject: { subject_code: 'Code03', acrostic: '', status: 'valid' }, site_id: sites(:two).id
+      post :create, project_id: @project, subject: { subject_code: 'Code03' }, site_id: sites(:two).id
     end
 
     assert_not_nil assigns(:project)
@@ -481,12 +481,12 @@ class SubjectsControllerTest < ActionController::TestCase
   end
 
   test 'should update subject' do
-    put :update, project_id: @project, id: @subject, subject: { subject_code: @subject.subject_code, acrostic: '', status: @subject.status }, site_id: @subject.site_id
+    put :update, project_id: @project, id: @subject, subject: { subject_code: @subject.subject_code }, site_id: @subject.site_id
     assert_redirected_to project_subject_path(assigns(:subject).project, assigns(:subject))
   end
 
   test 'should update subject with blank subject code' do
-    put :update, project_id: @project, id: @subject, subject: { subject_code: '', acrostic: '', status: @subject.status }, site_id: @subject.site_id
+    put :update, project_id: @project, id: @subject, subject: { subject_code: '' }, site_id: @subject.site_id
     assert_not_nil assigns(:subject)
     assert assigns(:subject).errors.size > 0
     assert_equal ["can't be blank"], assigns(:subject).errors[:subject_code]
@@ -494,13 +494,13 @@ class SubjectsControllerTest < ActionController::TestCase
   end
 
   test 'should not update invalid subject' do
-    put :update, id: -1, project_id: @project, subject: { subject_code: @subject.subject_code, acrostic: '', status: @subject.status }, site_id: @subject.site_id
+    put :update, id: -1, project_id: @project, subject: { subject_code: @subject.subject_code }, site_id: @subject.site_id
     assert_nil assigns(:subject)
     assert_redirected_to project_subjects_path
   end
 
   test 'should not update subject with invalid project' do
-    put :update, id: @subject, project_id: -1, subject: { subject_code: @subject.subject_code, acrostic: '', status: @subject.status }, site_id: @subject.site_id
+    put :update, id: @subject, project_id: -1, subject: { subject_code: @subject.subject_code }, site_id: @subject.site_id
 
     assert_nil assigns(:subject)
     assert_nil assigns(:project)
@@ -510,7 +510,7 @@ class SubjectsControllerTest < ActionController::TestCase
 
   test 'should not update subject for site editor for another site' do
     login(users(:site_one_editor))
-    put :update, id: subjects(:three), project_id: @project, subject: { subject_code: 'New Subject Code', acrostic: '', status: @subject.status }, site_id: sites(:one)
+    put :update, id: subjects(:three), project_id: @project, subject: { subject_code: 'New Subject Code' }, site_id: sites(:one)
 
     assert_not_nil assigns(:project)
     assert_nil assigns(:subject)